IntegerHashing
This python program is for integer hasing in which we take the array elements input using map() function which allow us to take input without knowing the number of elements in the list. Then we use a dictionary to count the occurrence of number in list . Dictionary in python is similar to map function in c++ which is used to store key value paris. Then we take the number of queries or number of elements for which we have to find occurrence and take query elements and print count.